+++ src/controllers/user.controller.ts
@@ -3,8 +3,6 @@ import { BaseError } from "../utils/errors/error";
 import { StatusCodes } from "http-status-codes";
 import { ChangePassword, TokenPayload, User, UserLogin, UserStatus } from "../domain/user";
 
-import jwt from "jsonwebtoken"
-
 import bcryptjs from "bcryptjs";
 import dotenv from "dotenv";
 import { log } from "console";
@@ -47,6 +45,7 @@ const create = async (user: User) => {
 
   const result =  await userDAO.create(user).catch(error => new BaseError("No se pudo registrar el usuario", StatusCodes.CONFLICT));
   if(result instanceof BaseError) throw result;
+  result.password='';
   return result;
 };
 
@@ -57,27 +56,26 @@ const del = async (id: string) => {
   return result;
 };
 
-const logIn = async (userLogin: UserLogin) => {
-  const userDAO = await new UserDAO();
+// const logIn = async (userLogin: UserLogin) => {
+//   const userDAO = await new UserDAO();
   
-  const result = await userDAO.findByUserName(userLogin.username) 
-                .catch((error: Error) => new BaseError(`El usuario ${userLogin.username} no esta registrado`, StatusCodes.NOT_FOUND, error.message));
+//   const result = await userDAO.findByUserName(userLogin.username) 
+//                 .catch((error: Error) => new BaseError(`El usuario ${userLogin.username} no esta registrado`, StatusCodes.NOT_FOUND, error.message));
 
-  if (!result || result instanceof BaseError) throw result; 
+//   if (!result || result instanceof BaseError) throw result; 
 
-  const isPasswordCorrect = await bcryptjs.compare(userLogin.password, result.password as string);
+//   const isPasswordCorrect = await bcryptjs.compare(userLogin.password, result.password as string);
 
-  if (!isPasswordCorrect) {
-    throw new BaseError('Contraseña incorrecta', StatusCodes.FORBIDDEN);
-  }
+//   if (!isPasswordCorrect) {
+//     throw new BaseError('Contraseña incorrecta', StatusCodes.FORBIDDEN);
+//   }
 
-  result.token = generateToken(result);
+//   result.token = generateToken(result);
 
-  const userWithToken = await userDAO.update(result.id, result)
-  userWithToken.password = '';
-  return userWithToken;
-// retornar usuario con un token
-}
+//   const userWithToken = await userDAO.update(result.id, result)
+//   userWithToken.password = '';
+//   return userWithToken;
+// }
 
 const update = async (id: string, item: User) => {
   const userDao = await new UserDAO();
@@ -87,17 +85,6 @@ const update = async (id: string, item: User) => {
   return result;
 }
 
-
-function generateToken(u: User): string {
-  if (!process.env.JWT_SECRET) {
-    throw new BaseError('Cannot generate token', StatusCodes.CONFLICT);
-  }
-  return jwt.sign(
-    { userId: u.id, role: u.role } as TokenPayload,
-    process.env.JWT_SECRET!,
-    { expiresIn: '1h' }
-  );
-}
 
 const changePassword = async (userName: string, item: ChangePassword) => {
   const userDao = await new UserDAO();
@@ -136,4 +123,4 @@ const changePassword = async (userName: string, item: ChangePassword) => {
 }
   
 
-export default { findOneById, getAll, create, delete: del, logIn, update, changePassword};
+export default { findOneById, getAll, create, delete: del, update, changePassword};
