+++ src/routes/user.routes.ts
@@ -1,9 +1,5 @@
 import { Router } from 'express';
 import {userHandler as handler} from '../handlers';
-import authMiddleware from '../middlewares';
-
-
-
 
 export default function usersRouter(): Router {
   return Router()
@@ -11,7 +7,7 @@ export default function usersRouter(): Router {
     .get('/:id', handler.findOneById)
     .get("/", handler.getAll)
     .post("/", handler.create)
-    .post("/login",handler.logIn)
+    //.post("/login",handler.logIn)
     .delete("/:id", handler.delete)
     .put("/changePassword", handler.changePassword)
     .put("/:id", handler.update)
